backend/pyproject.toml
[project]
name = "missing-table-backend"
version = "0.1.0"
description = "Sports league management backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.2",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "python-dateutil>=2.9.0.post0",
    "flask>=3.0.3",
    "flask-cors>=5.0.0",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "requests>=2.32.4",
    "httpcore>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "playwright>=1.53.0",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-instrumentation-requests>=0.49b0",
    "opentelemetry-instrumentation-redis>=0.49b0",
    "user-agents>=2.2.0",
    "celery>=5.5.3",
    "structlog>=25.4.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "jsonschema>=4.25.1",
    "kombu>=5.5.4",
    "crewai>=0.28.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "pyyaml>=6.0.2",
    "pytest-playwright>=0.7.1",
    "pytest-html>=4.1.1",
    "pillow>=12.0.0",
    "email-validator>=2.3.0",
    "debugpy>=1.8.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "coverage>=7.6.4",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "safety>=3.2.0",
    "pre-commit>=4.0.0",
    "types-requests>=2.32.0",
    "types-python-dateutil>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
mt-cli = "cli:app"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-xdist>=3.3.0",
    "coverage>=7.6.4",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "safety>=3.2.0",
    "pre-commit>=4.0.0",
    "types-requests>=2.32.0",
    "types-python-dateutil>=2.9.0",
    "httpx>=0.27.0",
    "schemathesis>=3.36.1",
    "openapi-spec-validator>=0.7.1",
    "datamodel-code-generator>=0.25.0",
    "respx>=0.21.0",
    "pytest-json-report>=1.5.0",
    "faker>=37.12.0",
    "allure-pytest>=2.13.0",
]

# Ruff configuration - Ultra-fast Python linter and formatter
[tool.ruff]
target-version = "py313"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "C90",  # mccabe complexity
    "S",    # bandit security
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # Ruff-specific rules
]
ignore = [
    "S101",   # Use of assert detected (OK for tests)
    "S608",   # Possible SQL injection (we'll handle with careful review)
    "B008",   # Do not perform function calls in argument defaults (OK for FastAPI)
    "ARG001", # Unused function argument (common in FastAPI for Depends)
    "ARG002", # Unused method argument (common in FastAPI)
    "ARG003", # Unused class method argument
    "S311",   # Standard pseudo-random generators (OK for non-crypto uses)
    "C901",   # Function too complex (refactoring deferred)
    "E402",   # Module level import not at top (often intentional)
    "SIM102", # Nested if (sometimes clearer with comments)
    "S110",   # try-except-pass (often intentional for suppression)
    "E741",   # Ambiguous variable name (l, O, I) - sometimes intentional
    "S105",   # Hardcoded password string (test defaults, config)
    "S106",   # Hardcoded password argument (test defaults, config)
    "S324",   # Insecure hash (md5 used for non-crypto purposes like cache keys)
    "S104",   # Binding to all interfaces (required for container deployment)
    "S112",   # try-except-continue (often intentional)
    "RUF012", # Mutable class attributes (Pydantic models)
    "SIM105", # contextlib.suppress (sometimes try/except is clearer)
    "SIM115", # context manager for files (sometimes not applicable)
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ARG", "E501"]
"**/migrations/*" = ["E501"]
"scripts/*" = ["S113", "S310", "S603", "S607", "T201", "E501"]  # Admin scripts: timeouts, subprocess, URL, print OK
"manage_*.py" = ["S113", "E501"]  # CLI tools: timeouts OK
"supabase-local/*" = ["S113", "E501"]  # Local setup scripts
"run_tests.py" = ["S603", "S607"]  # Test runner: subprocess OK

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy configuration - Static type checking
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "supabase.*",
    "pandas.*",
    "numpy.*",
    "playwright.*",
    "slowapi.*",
    "redis.*"
]
ignore_missing_imports = true

# Bandit configuration - Security linting
[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection for now

# Coverage configuration
[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2
fail_under = 80

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=. --cov-report=term-missing --cov-report=html"
pythonpath = ["..", "."]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    # Test Pyramid Levels (primary classification)
    "unit: Unit tests - pure logic, no dependencies",
    "integration: Integration tests - multiple components interacting",
    "component: Component tests - UI component testing (frontend)",
    "contract: API contract tests - validates API behavior",
    "e2e: End-to-end tests - full user workflows",

    # Stack Location
    "backend: Backend Python/FastAPI tests",
    "frontend: Frontend Vue/JavaScript tests",

    # Module/Feature Areas
    "auth: Authentication and authorization tests",
    "dao: Data access layer tests",
    "api: API endpoint tests",
    "invite: Invite system tests",
    "games: Game management tests",
    "standings: League standings tests",
    "admin: Admin functionality tests",

    # Performance Characteristics
    "fast: Fast tests < 100ms (default for unit tests)",
    "slow: Slow tests > 1 second (database, network)",

    # Dependencies
    "database: Requires database connection",
    "server: Requires running server",
    "mock: Uses mocks/stubs only",
    "external: Calls external services",

    # Other
    "debug: Debug tests for troubleshooting",
]
